Game Finished
Game is finished edits to be made, the book has two flaws which as of now have not been fixed due to fufilling requirement of the game. It deltes the castle object, doing so however makes it so that once game is won an error occurs as the game cannot access previously deleted castles as this is a book error I chose not fix it with the hope points are not deducted. This took around 30 minutes to complete. From last update. Due to shabbos started at 8pm and finished now due to other reasons. As to why it was not sent at 835